+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Fixed
 
 - CHANGELOG files are now copied to the build directory.
+- The `install=...` PKGBUILD field is respected.
 
 ## 4.1.0